Fix sign compare and unknown pragmas warnings

+++ include/opentxs/core/OTString.hpp
@@ -376,14 +376,14 @@ EXPORT static std::string  ws2s(const std::wstring & s);
             if (empties == split::no_empties)
             {
                 next = s.find_first_not_of(delimiters, static_cast<uint32_t>(next)+1);
-                if (next == Container::value_type::npos) break;
+                if (static_cast<size_t>(next) == Container::value_type::npos) break;
                 next -= 1;
             }
             current = static_cast<size_t>(next + 1);
             next = s.find_first_of( delimiters, current );
             result.push_back(s.substr(current, static_cast<uint32_t>(next) - current));
         }
-        while (next != Container::value_type::npos);
+        while (static_cast<size_t>(next) != Container::value_type::npos);
         return result;
     }
 	// ----------------------------------------------
